Rename overlay1.h -> overlay01.h
It needs to match the symbol table file name in order to be checked by
symbol_check.py
