+++ headers/functions/functions.h
@@ -2,7 +2,7 @@
 #define HEADERS_FUNCTIONS_H_
 
 #include "arm9.h"
-#include "overlay1.h"
+#include "overlay01.h"
 #include "overlay11.h"
 #include "overlay29.h"
 #include "overlay31.h"
